Split update checking UI code into UpdateManager
